requirements.txt
openai
elevenlabs
python-dotenv
requests
moviepy
imageio[ffmpeg]
pillow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
streamlit
